+++ react-rpoject/package.json
@@ -6,6 +6,7 @@
     "firebase": "^7.2.0",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
+    "react-router": "^5.1.2",
     "react-scripts": "3.2.0",
     "react-with-firebase-auth": "^1.1.0"
   },
