+++ react-rpoject/src/App.jsx
@@ -1,4 +1,7 @@
 import React from 'react';
+
+import Login from './components/Login';
+
 import withFirebaseAuth from 'react-with-firebase-auth';
 import * as firebase from 'firebase/app';
 import 'firebase/auth';
@@ -6,19 +9,16 @@ import firebaseConfig from './firebaseConfig.js';
 
 const firebaseApp = firebase.initializeApp(firebaseConfig);
 
-const firebaseAppAuth = firebaseApp.auth();
-const providers = {
-    googleProvider: new firebase.auth.GoogleAuthProvider(),
-};
-
-
-
 class App extends React.Component {
   render(){
     return(
       <header>
         <Title />
-        <Login />
+        <Login
+            user={this.props.user}
+            signInWithGoogle={this.props.signInWithGoogle}
+            signOut={this.props.signOut} 
+        />
       </header>
     )
   }
@@ -30,39 +30,13 @@ const Title = () => {
     )
 };
 
-class Login extends React.Component {
-    render(){
-        const {
-            user,
-            signOut,
-            signInWithGoogle,
-        } = this.props;
+const firebaseAppAuth = firebaseApp.auth();
 
-        return (
-            <div className="App">
-                <header className="App-header">
-                    
-                    {
-                        user
-                            ? <p>Hello, {user.displayName}</p>
-                            : <p>Please sign in.</p>
-                    }
-                    {
-                        user
-                            ? <button onClick={signOut}>Sign out</button>
-                            : <button onClick={signInWithGoogle}>Sign in with Google</button>
-                    }
-                </header>
-            </div>
-        );
-    }
+const providers = {
+    googleProvider: new firebase.auth.GoogleAuthProvider(),
 };
-
- 
-
 
 export default withFirebaseAuth({
     providers,
     firebaseAppAuth,
 })(App);
-
